requirements.txt
alembic==1.4.2
Babel==2.8.0
bcrypt==3.1.7
blinker==1.4
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
config==0.5.0.post0
cryptography==2.9.2
dnspython==2.0.0
email-validator==1.1.1
Flask==1.1.2
Flask-BabelEx==0.9.4
Flask-Bcrypt==0.7.1
Flask-JWT-Extended==3.24.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.4.4
Flask-User==0.6.21
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
passlib==1.7.2
Pillow==7.2.0
pkg-resources==0.0.0
pycparser==2.20
pycryptodome==3.9.8
PyJWT==1.7.1
PyMySQL==0.9.3
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pytz==2020.1
requests==2.24.0
six==1.15.0
speaklater==1.3
SQLAlchemy==1.3.18
urllib3==1.25.10
Werkzeug==1.0.1
WTForms==2.3.3
